cleanup `UseDifferentVisibility`
